refactored code, split dao to jdbc and hibernate
