+++ src/main/java/pl/hetman/wiktoria/java/app/socialactivitytracker/dao/hibernate/repository/UserActivityRepository.java
@@ -1,9 +1,8 @@
-package pl.hetman.wiktoria.java.app.socialactivitytracker.repositories;
+package pl.hetman.wiktoria.java.app.socialactivitytracker.dao.hibernate.repository;
 
 import org.hibernate.SessionFactory;
 import pl.hetman.wiktoria.java.app.socialactivitytracker.api.exception.ActivityException;
 import pl.hetman.wiktoria.java.app.socialactivitytracker.controller.model.ActivityModel;
-import pl.hetman.wiktoria.java.app.socialactivitytracker.dao.Dao;
 
 import java.util.List;
 import java.util.Optional;
