+++ src/main/java/pl/hetman/wiktoria/java/app/socialactivitytracker/services/ActivityService.java
@@ -2,8 +2,7 @@ package pl.hetman.wiktoria.java.app.socialactivitytracker.services;
 
 import pl.hetman.wiktoria.java.app.socialactivitytracker.api.exception.ActivityException;
 import pl.hetman.wiktoria.java.app.socialactivitytracker.controller.model.ActivityModel;
-import pl.hetman.wiktoria.java.app.socialactivitytracker.dao.ActivityDao;
-import pl.hetman.wiktoria.java.app.socialactivitytracker.repositories.ActivityRepository;
+import pl.hetman.wiktoria.java.app.socialactivitytracker.dao.hibernate.repository.ActivityRepository;
 
 import java.util.List;
 import java.util.Optional;
