+++ src/test/java/pl/hetman/wiktoria/java/app/socialactivitytracker/controller/ActivityControllerTest.java
@@ -5,8 +5,8 @@
 //import pl.hetman.wiktoria.java.app.socialactivitytracker.api.exception.ActivityException;
 //import pl.hetman.wiktoria.java.app.socialactivitytracker.api.exception.MissingLabelActivityException;
 //import pl.hetman.wiktoria.java.app.socialactivitytracker.controller.model.ActivityModel;
-//import pl.hetman.wiktoria.java.app.socialactivitytracker.dao.ActivityDao;
-//import pl.hetman.wiktoria.java.app.socialactivitytracker.dao.UserDao;
+//import pl.hetman.wiktoria.java.app.socialactivitytracker.dao.jdbc.ActivityDao;
+//import pl.hetman.wiktoria.java.app.socialactivitytracker.dao.jdbc.UserDao;
 //import pl.hetman.wiktoria.java.app.socialactivitytracker.service.ActivityService;
 //
 //import static org.junit.jupiter.api.Assertions.*;
