+++ src/test/java/pl/hetman/wiktoria/java/app/socialactivitytracker/dao/ActivityDaoListTest.java
@@ -4,10 +4,10 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import pl.hetman.wiktoria.java.app.socialactivitytracker.api.exception.ActivityException;
 import pl.hetman.wiktoria.java.app.socialactivitytracker.controller.model.ActivityModel;
+import pl.hetman.wiktoria.java.app.socialactivitytracker.dao.jdbc.ActivityDao;
+import pl.hetman.wiktoria.java.app.socialactivitytracker.dao.jdbc.UserDao;
 
 import java.util.List;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 class ActivityDaoListTest {
