+++ src/test/java/pl/hetman/wiktoria/java/app/socialactivitytracker/dao/ConnectionManagerTest.java
@@ -3,10 +3,9 @@ package pl.hetman.wiktoria.java.app.socialactivitytracker.dao;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import pl.hetman.wiktoria.java.app.socialactivitytracker.api.exception.ActivityException;
-import pl.hetman.wiktoria.java.app.socialactivitytracker.dao.ConnectionManager;
+import pl.hetman.wiktoria.java.app.socialactivitytracker.dao.jdbc.ConnectionManager;
 
 import java.sql.Connection;
-import java.sql.SQLException;
 
 class ConnectionManagerTest {
