+++ src/test/java/pl/hetman/wiktoria/java/app/socialactivitytracker/dao/UniqueIdGeneratorTest.java
@@ -2,7 +2,7 @@ package pl.hetman.wiktoria.java.app.socialactivitytracker.dao;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
-import pl.hetman.wiktoria.java.app.socialactivitytracker.dao.UniqueIdGenerator;
+import pl.hetman.wiktoria.java.app.socialactivitytracker.dao.jdbc.UniqueIdGenerator;
 
 class UniqueIdGeneratorTest {
